fix user controller

+++ src/Controller/UserController.php
@@ -106,17 +106,19 @@ class UserController extends BaseController
 
     }
 
-    #[Route('/{id}', name: 'app_user_show', methods: ['GET'])]
-    public function show(Request $request,User $user): Response
+    #[Route('/{id}', name: 'app_user_show', methods: ['GET'], requirements: ['id' => '\d+'])]
+    public function show(int $id, Request $request, UserRepository $userRepository): Response
     {
+        $user = $userRepository->findOneBy(['id'=>$id]);
         return $this->render('user/show.html.twig', [
             'user' => $user,
         ]);
     }
 
-    #[Route('/{id}/edit', name: 'app_user_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, User $user, UserRepository $userRepository): Response
+    #[Route('/{id}/edit', name: 'app_user_edit', methods: ['GET', 'POST'], requirements: ['id' => '\d+'])]
+    public function edit(int $id, Request $request, UserRepository $userRepository): Response
     {
+        $user = $userRepository->findOneBy(['id'=>$id]);
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
 
@@ -132,13 +134,14 @@ class UserController extends BaseController
         ]);
     }
 
-    #[Route('/{id}', name: 'app_user_delete', methods: ['POST'] )]
-    public function delete(Request $request, User $user, UserRepository $userRepository): Response
+    #[Route('/{id}', name: 'app_user_delete', methods: ['POST'], requirements: ['id' => '\d+'] )]
+    public function delete(int $id, Request $request, UserRepository $userRepository): Response
     {
-       /* if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
+        $user = $userRepository->findOneBy(['id'=>$id]);
+        if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
             $userRepository->remove($user, true);
         }
-*/
+
         return $this->redirectToRoute('app_user_index', [], Response::HTTP_SEE_OTHER);
     }
